Update form.t.js
Replace the field selectors with slightly improved approach that clicks all non-combo fields then clicks combos.  This is necessary in the case that the combo is not editable, which will cause it to expand when clicked.  If there are remaining fields that appear vertically positioned below the list, they are inaccessible to click.  This patch is not perfect because additional combos may be used sequentially on a form.